Improve the archetype-post-generate.groovy script by decomposing the code into 2 separate classes:  - Aether.groovy  - HelidonEngine.groovy
archetype-post-generate.groovy now only has bootstrap code, the main logic is in HelidonEngine.groovy.
It creates the class loader with all the maven core libs and uses a separate GroovyShell to create the instance of Aether.groovy
This removes the need to use reflection to bootstrap and use aether.

Because the main script is evaluate from a string there is no way to load the includes with a file, path or uri.
Thus, the 2 includes are inlined in the main script using multi line strings.