Fix aether utility to use Maven settings.
Fixes #383

Both the aether and engine utility are now written in Java.
- The byte code for the classes in included as base64 strings in the generated groovy script.
- Improve archetype-post-generate.groovy ; add comments
Update the integration-test mojo to generate the test projects with archetype:generate if mavenArchetypeCompatible==true.

Add a profile debug-invoker to allo debugging invoker based integration tests.
Add a settings.xml to enable better cache with the invoker plugin.