Use MojoExecutor instead of just BuildPluginManager.
The MavenProject object is stateful and is mutated before plugin executions.
Some of the stateful behavior has changed with newer Maven versions and it breaks the incremental compilation

While our previous implementation worked for simple usescases, there were very likely some plugins that would not
have executed properly since because we aren't mutating the state.

Using MojoExecutor which does seem to take care of all the stateful stuff.
However it is not a supported API of Maven and may change in the future, this is a trade-off.

Fixes #499